requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/dev.in
#
alembic==1.8.1
    # via -r requirements/base.in
anyio==3.6.1
    # via starlette
attrs==21.4.0
    # via flake8-bugbear
bandit==1.7.4
    # via -r requirements/dev.in
build==0.8.0
    # via pip-tools
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via
    #   pip-compile-multi
    #   pip-tools
    #   safety
    #   uvicorn
dparse==0.5.1
    # via safety
fastapi==0.79.0
    # via -r requirements/base.in
flake8==4.0.1
    # via
    #   -r requirements/dev.in
    #   flake8-bugbear
    #   flake8-builtins
    #   flake8-commas
    #   pep8-naming
flake8-bugbear==22.7.1
    # via -r requirements/dev.in
flake8-builtins==1.5.3
    # via -r requirements/dev.in
flake8-commas==2.1.0
    # via -r requirements/dev.in
flake8-import-order==0.18.1
    # via -r requirements/dev.in
flake8-variables-names==0.0.5
    # via -r requirements/dev.in
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via bandit
greenlet==1.1.2
    # via sqlalchemy
h11==0.13.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   requests
importlib-metadata==4.12.0
    # via alembic
importlib-resources==5.8.0
    # via alembic
mako==1.2.1
    # via alembic
markupsafe==2.1.1
    # via mako
mccabe==0.6.1
    # via flake8
packaging==21.3
    # via
    #   build
    #   dparse
    #   safety
pbr==5.9.0
    # via stevedore
pep517==0.12.0
    # via build
pep8-naming==0.13.0
    # via -r requirements/dev.in
pip-compile-multi==2.4.5
    # via -r requirements/base.in
pip-tools==6.8.0
    # via pip-compile-multi
psycopg2-binary==2.9.3
    # via -r requirements/base.in
pycodestyle==2.8.0
    # via
    #   flake8
    #   flake8-import-order
pydantic==1.9.1
    # via fastapi
pyflakes==2.4.0
    # via flake8
pyparsing==3.0.9
    # via packaging
python-dotenv==0.20.0
    # via -r requirements/base.in
pyyaml==6.0
    # via
    #   bandit
    #   dparse
requests==2.28.1
    # via safety
ruamel-yaml==0.17.21
    # via safety
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
safety==2.1.0
    # via -r requirements/dev.in
smmap==5.0.0
    # via gitdb
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.39
    # via
    #   -r requirements/base.in
    #   alembic
starlette==0.19.1
    # via fastapi
stevedore==4.0.0
    # via bandit
toml==0.10.2
    # via dparse
tomli==2.0.1
    # via
    #   build
    #   pep517
toposort==1.7
    # via pip-compile-multi
typing-extensions==4.3.0
    # via
    #   pydantic
    #   starlette
urllib3==1.26.10
    # via requests
uvicorn==0.18.2
    # via -r requirements/base.in
wheel==0.37.1
    # via pip-tools
zipp==3.8.1
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
